+++ src/Kerbalism/Automation/Computer.cs
@@ -1,25 +1,24 @@
-﻿using KSP.Localization;
+﻿using System.Linq;
 using System.Collections.Generic;
+using KSP.Localization;
 
 namespace KERBALISM
 {
-
-
 	public enum ScriptType
 	{
-		landed = 1,      // called on landing
-		atmo = 2,      // called on entering atmosphere
-		space = 3,      // called on reaching space
-		sunlight = 4,      // called when sun rise
-		shadow = 5,      // called when sun set
-		power_high = 6,      // called when ec level goes above 15%
-		power_low = 7,      // called when ec level goes below 15%
+		landed = 1,       // called on landing
+		atmo = 2,         // called on entering atmosphere
+		space = 3,        // called on reaching space
+		sunlight = 4,     // called when sun rise
+		shadow = 5,       // called when sun set
+		power_high = 6,   // called when ec level goes above 15%
+		power_low = 7,    // called when ec level goes below 15%
 		rad_low = 8,      // called when radiation goes below 0.05 rad/h
-		rad_high = 9,      // called when radiation goes above 0.05 rad/h
+		rad_high = 9,     // called when radiation goes above 0.05 rad/h
 		linked = 10,      // called when signal is regained
-		unlinked = 11,     // called when signal is lost
+		unlinked = 11,    // called when signal is lost
 		eva_out = 12,     // called when going out on eva
-		eva_in = 13,     // called when coming back from eva
+		eva_in = 13,      // called when coming back from eva
 		action1 = 14,     // called when pressing 1
 		action2 = 15,     // called when pressing 2
 		action3 = 16,     // called when pressing 3
@@ -28,7 +27,6 @@ namespace KERBALISM
 		last = 19
 	}
 
-
 	public sealed class Computer
 	{
 		public Computer()
@@ -36,74 +34,65 @@ namespace KERBALISM
 			scripts = new Dictionary<ScriptType, Script>();
 		}
 
-
-		public Computer( ConfigNode node )
+		public Computer(ConfigNode node)
 		{
 			// load scripts
 			scripts = new Dictionary<ScriptType, Script>();
-			foreach (var script_node in node.GetNode( "scripts" ).GetNodes())
+			foreach (var script_node in node.GetNode("scripts").GetNodes())
 			{
-				scripts.Add( (ScriptType) Lib.Parse.ToUInt( script_node.name ), new Script( script_node ) );
+				scripts.Add((ScriptType)Lib.Parse.ToUInt(script_node.name), new Script(script_node));
 			}
 		}
 
-
-		public void Save( ConfigNode node )
+		public void Save(ConfigNode node)
 		{
 			// save scripts
-			var scripts_node = node.AddNode( "scripts" );
+			var scripts_node = node.AddNode("scripts");
 			foreach (var p in scripts)
 			{
-				if (p.Value.states.Count == 0)
-					continue; //< empty-script optimization
-				p.Value.Save( scripts_node.AddNode( ((uint) p.Key).ToString() ) );
+				if (p.Value.states.Count == 0) continue; //< empty-script optimization
+				p.Value.Save(scripts_node.AddNode(((uint)p.Key).ToString()));
 			}
 		}
 
-
 		// get a script
-		public Script Get( ScriptType type )
+		public Script Get(ScriptType type)
 		{
-			if (!scripts.ContainsKey( type ))
-				scripts.Add( type, new Script() );
+			if (!scripts.ContainsKey(type)) scripts.Add(type, new Script());
 			return scripts[type];
 		}
 
-
 		// execute a script
-		public void Execute( Vessel v, ScriptType type )
+		public void Execute(Vessel v, ScriptType type)
 		{
 			// do nothing if there is no EC left on the vessel
-			Resource_info ec = ResourceCache.Info( v, "ElectricCharge" );
-			if (ec.amount <= double.Epsilon)
-				return;
+			Resource_info ec = ResourceCache.Info(v, "ElectricCharge");
+			if (ec.amount <= double.Epsilon) return;
 
 			// get the script
 			Script script;
-			if (scripts.TryGetValue( type, out script ))
+			if (scripts.TryGetValue(type, out script))
 			{
 				// execute the script
-				script.Execute( Boot( v ) );
+				script.Execute(Boot(v));
 
 				// show message to the user
 				// - unless the script is empty (can happen when being edited)
-				if (script.states.Count > 0 && DB.Vessel( v ).cfg_script)
+				if (script.states.Count > 0 && DB.Vessel(v).cfg_script)
 				{
-					Message.Post( Lib.BuildString( Localizer.Format( "#KERBALISM_UI_scriptvessel" ), " <b>", v.vesselName, "</b>" ) );
+					Message.Post(Lib.BuildString(Localizer.Format("#KERBALISM_UI_scriptvessel"), " <b>", v.vesselName, "</b>"));
 				}
 			}
 		}
 
-
 		// call scripts automatically when conditions are met
-		public void Automate( Vessel v, Vessel_info vi, Vessel_resources resources )
+		public void Automate(Vessel v, Vessel_info vi, Vessel_resources resources)
 		{
 			// do nothing if automation is disabled
-			if (!Features.Automation)
-				return;
+			if (!Features.Automation) return;
 
 			// get current states
-			Resource_info ec = resources.Info( v, "ElectricCharge" );
+			Resource_info ec = resources.Info(v, "ElectricCharge");
 			bool sunlight = vi.sunlight > double.Epsilon;
 			bool power_low = ec.level < 0.2;
 			bool power_high = ec.level > 0.8;
@@ -140,74 +129,62 @@ namespace KERBALISM
 			{
 				ScriptType type = p.Key;
 				Script script = p.Value;
-				if (script.states.Count == 0)
-					continue; //< skip empty scripts (may happen during editing)
+				if (script.states.Count == 0) continue; //< skip empty scripts (may happen during editing)
 
 				switch (type)
 				{
 					case ScriptType.landed:
-						if (landed && script.prev == "0")
-							to_exec.Add( script );
+						if (landed && script.prev == "0") to_exec.Add(script);
 						script.prev = landed ? "1" : "0";
 						break;
 
 					case ScriptType.atmo:
-						if (atmo && script.prev == "0")
-							to_exec.Add( script );
+						if (atmo && script.prev == "0") to_exec.Add(script);
 						script.prev = atmo ? "1" : "0";
 						break;
 
 					case ScriptType.space:
-						if (space && script.prev == "0")
-							to_exec.Add( script );
+						if (space && script.prev == "0") to_exec.Add(script);
 						script.prev = space ? "1" : "0";
 						break;
 
 					case ScriptType.sunlight:
-						if (sunlight && script.prev == "0")
-							to_exec.Add( script );
+						if (sunlight && script.prev == "0") to_exec.Add(script);
 						script.prev = sunlight ? "1" : "0";
 						break;
 
 					case ScriptType.shadow:
-						if (!sunlight && script.prev == "0")
-							to_exec.Add( script );
+						if (!sunlight && script.prev == "0") to_exec.Add(script);
 						script.prev = !sunlight ? "1" : "0";
 						break;
 
 					case ScriptType.power_high:
-						if (power_high && script.prev == "0")
-							to_exec.Add( script );
+						if (power_high && script.prev == "0") to_exec.Add(script);
 						script.prev = power_high ? "1" : "0";
 						break;
 
 					case ScriptType.power_low:
-						if (power_low && script.prev == "0")
-							to_exec.Add( script );
+						if (power_low && script.prev == "0") to_exec.Add(script);
 						script.prev = power_low ? "1" : "0";
 						break;
 
 					case ScriptType.rad_low:
-						if (radiation_low && script.prev == "0")
-							to_exec.Add( script );
+						if (radiation_low && script.prev == "0") to_exec.Add(script);
 						script.prev = radiation_low ? "1" : "0";
 						break;
 
 					case ScriptType.rad_high:
-						if (radiation_high && script.prev == "0")
-							to_exec.Add( script );
+						if (radiation_high && script.prev == "0") to_exec.Add(script);
 						script.prev = radiation_high ? "1" : "0";
 						break;
 
 					case ScriptType.linked:
-						if (signal && script.prev == "0")
-							to_exec.Add( script );
+						if (signal && script.prev == "0") to_exec.Add(script);
 						script.prev = signal ? "1" : "0";
 						break;
 
 					case ScriptType.unlinked:
-						if (!signal && script.prev == "0")
-							to_exec.Add( script );
+						if (!signal && script.prev == "0") to_exec.Add(script);
 						script.prev = !signal ? "1" : "0";
 						break;
 				}
@@ -218,26 +195,25 @@ namespace KERBALISM
 			{
 				// get list of devices
 				// - we avoid creating it when there are no scripts to be executed, making its overall cost trivial
-				var devices = Boot( v );
+				var devices = Boot(v);
 
 				// execute all scripts
 				foreach (Script script in to_exec)
 				{
-					script.Execute( devices );
+					script.Execute(devices);
 				}
 
 				// show message to the user
-				if (DB.Vessel( v ).cfg_script)
+				if (DB.Vessel(v).cfg_script)
 				{
-					Message.Post( Lib.BuildString( "Script called on vessel <b>", v.vesselName, "</b>" ) );
+					Message.Post(Lib.BuildString("Script called on vessel <b>", v.vesselName, "</b>"));
 				}
 			}
 		}
 
-
 		// return set of devices on a vessel
 		// - the list is only valid for a single simulation step
-		public static Dictionary<uint, Device> Boot( Vessel v )
+		public static Dictionary<uint, Device> Boot(Vessel v)
 		{
 			// store all devices
 			var devices = new Dictionary<uint, Device>();
@@ -248,79 +224,38 @@ namespace KERBALISM
 			// loaded vessel
 			if (v.loaded)
 			{
-				foreach (PartModule m in Lib.FindModules<PartModule>( v ))
+				foreach (PartModule m in Lib.FindModules<PartModule>(v))
 				{
 					switch (m.moduleName)
 					{
-						case "ProcessController":
-							dev = new ProcessDevice( m as ProcessController );
-							break;
-						case "Greenhouse":
-							dev = new GreenhouseDevice( m as Greenhouse );
-							break;
-						case "GravityRing":
-							dev = new RingDevice( m as GravityRing );
-							break;
-						case "Emitter":
-							dev = new EmitterDevice( m as Emitter );
-							break;
-						case "Laboratory":
-							dev = new LaboratoryDevice( m as Laboratory );
-							break;
-						case "Experiment":
-							dev = new ExperimentDevice( m as Experiment );
-							break;
-						case "ModuleDeployableSolarPanel":
-							dev = new PanelDevice( m as ModuleDeployableSolarPanel );
-							break;
-						case "ModuleGenerator":
-							dev = new GeneratorDevice( m as ModuleGenerator );
-							break;
-						case "ModuleResourceConverter":
-							dev = new ConverterDevice( m as ModuleResourceConverter );
-							break;
-						case "ModuleKPBSConverter":
-							dev = new ConverterDevice( m as ModuleResourceConverter );
-							break;
-						case "FissionReactor":
-							dev = new ConverterDevice( m as ModuleResourceConverter );
-							break;
-						case "ModuleResourceHarvester":
-							dev = new DrillDevice( m as ModuleResourceHarvester );
-							break;
-						case "ModuleLight":
-							dev = new LightDevice( m as ModuleLight );
-							break;
-						case "ModuleColoredLensLight":
-							dev = new LightDevice( m as ModuleLight );
-							break;
-						case "ModuleMultiPointSurfaceLight":
-							dev = new LightDevice( m as ModuleLight );
-							break;
-						case "SCANsat":
-							dev = new ScannerDevice( m );
-							break;
-						case "ModuleSCANresourceScanner":
-							dev = new ScannerDevice( m );
-							break;
-						case "ModuleRTAntenna":
-							dev = new RemoteTechAntennaDevice( m );
-							break;
-						case "ModuleRTAntennaPassive":
-							dev = new RemoteTechAntennaDevice( m );
-							break;
-						case "ModuleDeployableAntenna":
-							dev = new DeployableAntennaDevice( m as ModuleDeployableAntenna );
-							break;
-						default:
-							continue;
+						case "ProcessController":            dev = new ProcessDevice(m as ProcessController);                 break;
+						case "Greenhouse":                   dev = new GreenhouseDevice(m as Greenhouse);                     break;
+						case "GravityRing":                  dev = new RingDevice(m as GravityRing);                          break;
+						case "Emitter":                      dev = new EmitterDevice(m as Emitter);                           break;
+						case "Laboratory":                   dev = new LaboratoryDevice(m as Laboratory);                     break;
+						case "Experiment":                   dev = new ExperimentDevice(m as Experiment);                     break;
+						case "ModuleDeployableSolarPanel":   dev = new PanelDevice(m as ModuleDeployableSolarPanel);          break;
+						case "ModuleGenerator":              dev = new GeneratorDevice(m as ModuleGenerator);                 break;
+						case "ModuleResourceConverter":      dev = new ConverterDevice(m as ModuleResourceConverter);         break;
+						case "ModuleKPBSConverter":          dev = new ConverterDevice(m as ModuleResourceConverter);         break;
+						case "FissionReactor":               dev = new ConverterDevice(m as ModuleResourceConverter);         break;
+						case "ModuleResourceHarvester":      dev = new DrillDevice(m as ModuleResourceHarvester);             break;
+						case "ModuleLight":                  dev = new LightDevice(m as ModuleLight);                         break;
+						case "ModuleColoredLensLight":       dev = new LightDevice(m as ModuleLight);                         break;
+						case "ModuleMultiPointSurfaceLight": dev = new LightDevice(m as ModuleLight);                         break;
+						case "SCANsat":                      dev = new ScannerDevice(m);                                      break;
+						case "ModuleSCANresourceScanner":    dev = new ScannerDevice(m);                                      break;
+						case "ModuleRTAntenna":              dev = new Antenna(m, m.moduleName);                              break;
+						case "ModuleRTAntennaPassive":       dev = new Antenna(m, "ModuleRTAntenna");                         break;
+						case "ModuleDataTransmitter":        dev = new Antenna(m, m.moduleName);                              break;
+						default: continue;
 					}
 
 					// add the device
 					// - multiple same-type components in the same part will have the same id, and are ignored
-					if (!devices.ContainsKey( dev.Id() ))
+					if (!devices.ContainsKey(dev.Id()))
 					{
-						devices.Add( dev.Id(), dev );
+						devices.Add(dev.Id(), dev);
 					}
 				}
 			}
@@ -334,7 +269,7 @@ namespace KERBALISM
 				foreach (ProtoPartSnapshot p in v.protoVessel.protoPartSnapshots)
 				{
 					// get part prefab (required for module properties)
-					Part part_prefab = PartLoader.getPartInfoByName( p.partName ).partPrefab;
+					Part part_prefab = PartLoader.getPartInfoByName(p.partName).partPrefab;
 
 					// get all module prefabs
 					var module_prefabs = part_prefab.FindModulesImplementing<PartModule>();
@@ -347,100 +282,54 @@ namespace KERBALISM
 					{
 						// get the module prefab
 						// if the prefab doesn't contain this module, skip it
-						PartModule module_prefab = Lib.ModulePrefab( module_prefabs, m.moduleName, PD );
-						if (!module_prefab)
-							continue;
+						PartModule module_prefab = Lib.ModulePrefab(module_prefabs, m.moduleName, PD);
+						if (!module_prefab) continue;
 
 						// if the module is disabled, skip it
 						// note: this must be done after ModulePrefab is called, so that indexes are right
-						if (!Lib.Proto.GetBool( m, "isEnabled" ))
-							continue;
+						if (!Lib.Proto.GetBool(m, "isEnabled")) continue;
 
 						// depending on module name
 						switch (m.moduleName)
 						{
-							case "ProcessController":
-								dev = new ProtoProcessDevice( m, module_prefab as ProcessController, p.flightID );
-								break;
-							case "Greenhouse":
-								dev = new ProtoGreenhouseDevice( m, p.flightID );
-								break;
-							case "GravityRing":
-								dev = new ProtoRingDevice( m, p.flightID );
-								break;
-							case "Emitter":
-								dev = new ProtoEmitterDevice( m, p.flightID );
-								break;
-							case "Laboratory":
-								dev = new ProtoLaboratoryDevice( m, p.flightID );
-								break;
-							case "Experiment":
-								dev = new ProtoExperimentDevice( m, module_prefab as Experiment, p.flightID );
-								break;
-							case "ModuleDeployableSolarPanel":
-								dev = new ProtoPanelDevice( m, module_prefab as ModuleDeployableSolarPanel, p.flightID );
-								break;
-							case "ModuleGenerator":
-								dev = new ProtoGeneratorDevice( m, module_prefab as ModuleGenerator, p.flightID );
-								break;
-							case "ModuleResourceConverter":
-								dev = new ProtoConverterDevice( m, module_prefab as ModuleResourceConverter, p.flightID );
-								break;
-							case "ModuleKPBSConverter":
-								dev = new ProtoConverterDevice( m, module_prefab as ModuleResourceConverter, p.flightID );
-								break;
-							case "FissionReactor":
-								dev = new ProtoConverterDevice( m, module_prefab as ModuleResourceConverter, p.flightID );
-								break;
-							case "ModuleResourceHarvester":
-								dev = new ProtoDrillDevice( m, module_prefab as ModuleResourceHarvester, p.flightID );
-								break;
-							case "ModuleLight":
-								dev = new ProtoLightDevice( m, p.flightID );
-								break;
-							case "ModuleColoredLensLight":
-								dev = new ProtoLightDevice( m, p.flightID );
-								break;
-							case "ModuleMultiPointSurfaceLight":
-								dev = new ProtoLightDevice( m, p.flightID );
-								break;
-							case "SCANsat":
-								dev = new ProtoScannerDevice( m, part_prefab, v, p.flightID );
-								break;
-							case "ModuleSCANresourceScanner":
-								dev = new ProtoScannerDevice( m, part_prefab, v, p.flightID );
-								break;
-							case "ModuleRTAntenna":
-								dev = new ProtoRemoteTechAntennaDevice( m, module_prefab, v, p.flightID );
-								break;
-							case "ModuleRTAntennaPassive":
-								dev = new ProtoRemoteTechAntennaDevice( m, module_prefab, v, p.flightID );
-								break;
-							case "ModuleDeployableAntenna":
-								dev = new ProtoDeployableAntennaDevice( m, module_prefab as ModuleDeployableAntenna, p.flightID );
-								break;
-							default:
-								continue;
+							case "ProcessController":            dev = new ProtoProcessDevice(m, module_prefab as ProcessController, p.flightID);                 break;
+							case "Greenhouse":                   dev = new ProtoGreenhouseDevice(m, p.flightID);                                                  break;
+							case "GravityRing":                  dev = new ProtoRingDevice(m, p.flightID);                                                        break;
+							case "Emitter":                      dev = new ProtoEmitterDevice(m, p.flightID);                                                     break;
+							case "Laboratory":                   dev = new ProtoLaboratoryDevice(m, p.flightID);                                                  break;
+							case "Experiment":                   dev = new ProtoExperimentDevice(m, module_prefab as Experiment, p.flightID);                     break;
+							case "ModuleDeployableSolarPanel":   dev = new ProtoPanelDevice(m, module_prefab as ModuleDeployableSolarPanel, p.flightID);          break;
+							case "ModuleGenerator":              dev = new ProtoGeneratorDevice(m, module_prefab as ModuleGenerator, p.flightID);                 break;
+							case "ModuleResourceConverter":      dev = new ProtoConverterDevice(m, module_prefab as ModuleResourceConverter, p.flightID);         break;
+							case "ModuleKPBSConverter":          dev = new ProtoConverterDevice(m, module_prefab as ModuleResourceConverter, p.flightID);         break;
+							case "FissionReactor":               dev = new ProtoConverterDevice(m, module_prefab as ModuleResourceConverter, p.flightID);         break;
+							case "ModuleResourceHarvester":      dev = new ProtoDrillDevice(m, module_prefab as ModuleResourceHarvester, p.flightID);             break;
+							case "ModuleLight":                  dev = new ProtoLightDevice(m, p.flightID);                                                       break;
+							case "ModuleColoredLensLight":       dev = new ProtoLightDevice(m, p.flightID);                                                       break;
+							case "ModuleMultiPointSurfaceLight": dev = new ProtoLightDevice(m, p.flightID);                                                       break;
+							case "SCANsat":                      dev = new ProtoScannerDevice(m, part_prefab, v, p.flightID);                                     break;
+							case "ModuleSCANresourceScanner":    dev = new ProtoScannerDevice(m, part_prefab, v, p.flightID);                                     break;
+							case "ModuleRTAntenna":              dev = new ProtoPartAntenna(m, p, v, m.moduleName, p.flightID);                                   break;
+							case "ModuleRTAntennaPassive":       dev = new ProtoPartAntenna(m, p, v, "ModuleRTAntenna", p.flightID);                              break;
+							case "ModuleDataTransmitter":        dev = new ProtoPartAntenna(m, p, v, m.moduleName, p.flightID);                                   break;
+							default: continue;
 						}
 
 						// add the device
 						// - multiple same-type components in the same part will have the same id, and are ignored
-						if (!devices.ContainsKey( dev.Id() ))
+						if (!devices.ContainsKey(dev.Id()))
 						{
-							devices.Add( dev.Id(), dev );
+							devices.Add(dev.Id(), dev);
 						}
 					}
 				}
 			}
 
-			// return all devices found
+			devices = devices.OrderBy(k => k.Value.Name()).ToDictionary(pair => pair.Key, pair => pair.Value);
+			//return all found devices sorted by name
 			return devices;
 		}
 
-
 		Dictionary<ScriptType, Script> scripts;
 	}
-
-
 } // KERBALISM
-